% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon.R
\docType{data}
\name{argon_history}
\alias{argon_history}
\title{Canonical coordinates of argon dataset}
\format{
An object of class \code{data.frame} with 25000 rows and 4 columns.
}
\usage{
argon_history
}
\description{
canonical variables data of 250 argon atoms at 94 K and 1 atm in an Nosé-Hoover
NPT ensemble
}
\keyword{datasets}
