% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon.R
\docType{data}
\name{argon_thermo}
\alias{argon_thermo}
\title{Thermodynamic argon dataset}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\usage{
argon_thermo
}
\arguments{
\item{time}{Time}

\item{volume}{Volume}

\item{epotential}{Potential energy}

\item{ekinetic}{Kinetic energy}

\item{temperature}{Temperature}

\item{pressure}{Pressure}
}
\description{
Thermodynamics data of 250 argon atoms at 94 K and 1 atm in an Nosé-Hoover
NPT ensemble
}
\keyword{datasets}
